Allow for simple re-usable `AddPropertyVisitor`
